refactored logging to become a plugin
